Remove Unecessary Comment
Multidex is used because we use more dependency and APK files contain executable bytecode files in the form of Dalvik Executable (DEX) files.  The DEX specification limits the total number of methods that can be referenced within a single DEX file to 65,536 (64K reference limit). We reference more methods than the number then we need to use multidex.